+++ src/Sutil/Html.fs
@@ -8,9 +8,8 @@ open System
 // Dummy type to avoid problems with overload resolution in HtmlEngine
 type [<Fable.Core.Erase>] NodeAttr = NodeAttr of SutilElement
 
-type SutilEventEngine() as this =
+type SutilEventEngine() =
     inherit EventEngine<SutilElement>( fun (event:string) handler -> Sutil.Attr.on (event.ToLower()) handler [] )
-
 
 type SutilHtmlEngine() as this =
     inherit HtmlEngine<SutilElement>( el, text, (fun () -> fragment []) )
@@ -81,6 +80,7 @@ type SutilAttrEngine() =
             bindAttrBoth "value" value dispatch
 
     member _.style (cssAttrs : (string*obj) seq) = cssAttrs |> Sutil.Attr.style
+    member _.styleAppend (cssAttrs : (string*obj) seq) = cssAttrs |> Sutil.Attr.styleAppend
     member _.style (cssAttrs : IObservable< #seq<string*obj> >) = Bind.style cssAttrs
 
     member _.none = nodeFactory <| fun ctx -> unitResult(ctx,"none")
@@ -98,30 +98,11 @@ let cssAttr = id
 let addClass       (n:obj) = cssAttr("sutil-add-class",n)
 let useGlobal              = cssAttr("sutil-use-global","" :> obj)
 
+type Media() =
+    static member Custom (condition : string) rules = makeMediaRule condition rules
+    static member MinWidth (minWidth : Styles.ICssUnit, rules : StyleSheetDefinition list) = makeMediaRule (sprintf "(min-width: %s)" (string minWidth)) rules
+    static member MaxWidth (maxWidth : Styles.ICssUnit, rules : StyleSheetDefinition list) = makeMediaRule (sprintf "(max-width: %s)" (string maxWidth)) rules
+
 // Convenience
 let text s = DOM.text s
 
-let exampleVirtualNodes =
-    Html.div [
-        text "Hello"
-        fragment [
-            text "World"
-        ]
-        fragment [
-            text "A"
-            fragment [
-                text "B"
-            ]
-            text "C"
-        ]
-    ]
-
-(*
-
-   DIVElement
-      TextNode "Hello"
-      TextNode "World"
-      TextNode "A"
-      TextNode "B"
-      TextNode "C"
-*)